feat: a example for the fixing of call api routes from frontend
